+++ test/cloregram/test_handlers.clj
@@ -1,9 +1,10 @@
 (ns cloregram.test-handlers
   (:require [cloregram.api :as api]
-            [cloregram.utils :as utl]
-            [clojure.string :as str]))
+            [clojure.string :as str]
+            [dialog.logger :as log]
+            [cloregram.utils :as utl]))
 
-(defn common
+(defn main-handler
   [{:keys [user message]}]
   (api/send-message user
                     (str (:user/username user) " " (str/upper-case (:text message)))
@@ -20,8 +21,31 @@
   [{:keys [n user]}]
   (let [n (dec n)]
     (api/send-message user (format "Decremented: %d" n)
-                    [[["+" 'cloregram.test-handlers/increment {:n n}]["-" 'cloregram.test-handlers/decrement {:n n}]]])))
+                      [[["+" 'cloregram.test-handlers/increment {:n n}]["-" 'cloregram.test-handlers/decrement {:n n}]]])))
+
+(defonce ^:private temp-id (atom nil))
 
 (defn temp
   [{:keys [user]}]
-  (api/send-message user "Temp message" [] :temp))
+  (->> (api/send-message user "Temp message" [[["New text" 'cloregram.test-handlers/new-temp]]] :temp)
+       :message_id
+       (reset! temp-id)))
+
+(defn new-temp
+  [{:keys [user]}]
+  (api/send-message user "New temp message" [[["New text 2" 'cloregram.test-handlers/new-temp-2]]] :temp @temp-id))
+
+(defn new-temp-2
+  [{:keys [user]}]
+  (log/debug "<=====")
+  (api/send-message user "New temp message 2" nil :temp @temp-id))
+
+(defn photo-handler
+  [{user :user {:keys [photo]} :message}]
+  (let [photo# (fn [u pss]
+                 (let [ps (max-key :width pss)
+                       file-id (:file_id ps)]
+                   ))]
+    (if photo
+      (photo# user photo)
+      (api/send-message user "Image expected!" [] :temp))))